+++ ntex-net/src/tokio/mod.rs
@@ -1,17 +1,21 @@
 #![allow(unused_variables)]
 #[cfg(unix)]
 use std::os::unix::net::UnixStream as OsUnixStream;
+use std::{io::Result as IoResult, net, net::SocketAddr, path::PathBuf};
 
 use ntex_io::Io;
 use ntex_service::cfg::SharedCfg;
 
-mod io_impl;
-
-pub use self::io_impl::TokioIoBoxed;
-use crate::channel::{self, Receiver};
-
 #[doc(hidden)]
-pub use tok_io::*;
+pub mod internal {
+    pub use tok_io::*;
+}
+
+use crate::channel;
+
+mod io;
+
+pub use self::io::TokioIoBoxed;
 
 pub(crate) struct TcpStream(tok_io::net::TcpStream);
 
@@ -51,7 +55,7 @@ impl ntex_rt::Driver for TokioDriver {
 }
 
 impl crate::Reactor for TokioDriver {
-    fn tcp_connect(&self, addr: std::net::SocketAddr, cfg: SharedCfg) -> Receiver<Io> {
+    fn tcp_connect(&self, addr: SocketAddr, cfg: SharedCfg) -> channel::Receiver<Io> {
         let (tx, rx) = channel::create();
         ntex_rt::spawn(async move {
             let result = async {
@@ -66,7 +70,7 @@ impl crate::Reactor for TokioDriver {
         rx
     }
 
-    fn unix_connect(&self, addr: std::path::PathBuf, cfg: SharedCfg) -> Receiver<Io> {
+    fn unix_connect(&self, addr: PathBuf, cfg: SharedCfg) -> channel::Receiver<Io> {
         #[cfg(unix)]
         {
             let (tx, rx) = channel::create();
@@ -84,17 +88,13 @@ impl crate::Reactor for TokioDriver {
 
         #[cfg(not(unix))]
         {
-            Receiver::new(Err(std::io::Error::other(
+            crate::channel::Receiver::new(Err(std::io::Error::other(
                 "Unix domain sockets are not supported",
             )))
         }
     }
 
-    fn from_tcp_stream(
-        &self,
-        stream: std::net::TcpStream,
-        cfg: SharedCfg,
-    ) -> std::io::Result<Io> {
+    fn from_tcp_stream(&self, stream: net::TcpStream, cfg: SharedCfg) -> IoResult<Io> {
         stream.set_nonblocking(true)?;
         stream.set_nodelay(true)?;
         Ok(Io::new(
@@ -104,11 +104,7 @@ impl crate::Reactor for TokioDriver {
     }
 
     #[cfg(unix)]
-    fn from_unix_stream(
-        &self,
-        stream: OsUnixStream,
-        cfg: SharedCfg,
-    ) -> std::io::Result<Io> {
+    fn from_unix_stream(&self, stream: OsUnixStream, cfg: SharedCfg) -> IoResult<Io> {
         stream.set_nonblocking(true)?;
         Ok(Io::new(
             UnixStream(tok_io::net::UnixStream::from_std(stream)?),